requirements.txt
torch
colorama
APScheduler
black
click
datasets
gradio
gradio_client
huggingface-hub
matplotlib
numpy
pandas
plotly
python-dateutil
requests
semantic-version
tqdm
wandb
transformers>=4.36.0
tokenizers>=0.15.0
lm_eval[ifeval] @ git+https://github.com/EleutherAI/lm-evaluation-harness.git@v0.4.2
accelerate
sentencepiece
langdetect
sacrebleu
cchardet
rouge_score
bert-score
evaluate
spacy==3.7.4
selfcheckgpt
immutabledict
gputil
bitsandbytes